Fix Warning: Empty `slug` generated for ''
We have a blog post with an empty category, which is being used in the
following code of the template:
```
{ category_name | slugify }
```

This causes the following warning:
```
Warning: Empty `slug` generated for ''.
```

Even more important, it renders an empty category in the menu.

Remove the empty category to fix this issue.
We may want to change how this is implemented in the template to avoid
problems with empty categories.

Closes https://github.com/openSUSE/news-o-o/issues/46