Fix Jemoji integration caused by too general css
IMO we shouldn't use display block for all images inside an article. This breaks
plugins (such us jemoji) and doesn't allow the authors to decide how they
want to render the images in their blog posts (for example if they want to
render three images in a row).

As I am not sure why this is set and where this is used and it could be a
breakable change, keep the display block for all images by now. Add a
TODO comment and fix it only for images with the `emoji` class (the ones
generated by the plugin).