Fix my blog posts and add license & image
Some things got broken in my blog posts:
- Emojis: Fix them using Jemoji plugin
- Extra ###: Remove them
- Broken links: Fix them by adding the missed `https:` (only rendering
was broken)
- Deleted embedded video: Add it back

Some of these problems may be present in other blog posts as well, but there are
too many to check all of them.

I have also added the license to the subtitle (it was already at the end
of the blog posts) and a header image to all my blog posts.